+++ pages/notes/home.vue
@@ -33,29 +33,35 @@ definePageMeta({
 })
 
 let { data: databaseNotes } = await useAsyncData('databaseNotes', async () => {
-  if (user) {
+  if (user.value) {
     const { data } = await client.from('notes').select('id, title, body').eq('user_id', user.value.id).limit(4)
     return data
   }
 })
 
 async function addNote(newNoteTitle: string, newNoteTag: string, newNoteBody: string) {
-  const { data } = await client
-    .from('notes')
-    .insert([
-      { title: newNoteTitle, tag: newNoteTag, body: newNoteBody, user_id: user.value.id },
-    ])
-    .select('id, title, tag, body')
-  databaseNotes.value.push(data[0])
+  if (user.value && databaseNotes.value) {
+    const { data } = await client
+      .from('notes')
+      .insert([
+        { title: newNoteTitle, tag: newNoteTag, body: newNoteBody, user_id: user.value.id },
+      ])
+      .select('id, title, tag, body')
+    if (data) {
+      databaseNotes.value.push(data[0])
+    }
+  }
 }
 
 async function deleteNote(id: number) {
-  await client
-    .from('notes')
-    .delete()
-    .eq('id', id)
-  let newDatabaseNotes = databaseNotes.filter(i => i.id != id)
-  databaseNotes.value = newDatabaseNotes
+  if (databaseNotes.value) {
+    await client
+      .from('notes')
+      .delete()
+      .eq('id', id)
+    let newDatabaseNotes = databaseNotes.value.filter((i: { id: number; title: string; body: string }) => i.id != id)
+    databaseNotes.value = newDatabaseNotes
+  }
 }
 async function filterNotes(id: number) {
