+++ pages/notes/my-notes.vue
@@ -16,7 +16,7 @@ definePageMeta({
 })
 
 let { data: databaseNotes } = await useAsyncData('databaseNotes', async () => {
-  if (user) {
+  if (user.value) {
     const { data } = await client.from('notes').select('id, title, body').eq('user_id', user.value.id)
     if (data) {
       return data
@@ -26,21 +26,27 @@ let { data: databaseNotes } = await useAsyncData('databaseNotes', async () => {
 })
 
 async function addNote(newNoteTitle: string, newNoteTag: string, newNoteBody: string) {
-  const { data } = await client
-    .from('notes')
-    .insert([
-      { title: newNoteTitle, tag: newNoteTag, body: newNoteBody, user_id: user.value.id },
-    ])
-    .select('id, title, tag, body')
-  databaseNotes.value.push(data[0])
+  if (user.value && databaseNotes.value) {
+    const { data } = await client
+      .from('notes')
+      .insert([
+        { title: newNoteTitle, tag: newNoteTag, body: newNoteBody, user_id: user.value.id },
+      ])
+      .select('id, title, tag, body')
+    if (data) {
+      databaseNotes.value.push(data[0])
+    }
+  }
 }
 
 async function deleteNote(id: number) {
-  await client
-    .from('notes')
-    .delete()
-    .eq('id', id)
-  let newDatabaseNotes = databaseNotes.value.filter((i: { id: number }) => i.id != id)
-  databaseNotes.value = newDatabaseNotes
+  if (databaseNotes.value) {
+    await client
+      .from('notes')
+      .delete()
+      .eq('id', id)
+    let newDatabaseNotes = databaseNotes.value.filter((i: { id: number; title: string; body: string }) => i.id != id)
+    databaseNotes.value = newDatabaseNotes
+  }
 }
 </script>
